backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
azure-cognitiveservices-speech==1.31.0
PyPDF2==3.0.1
python-docx==0.8.11
SQLAlchemy==2.0.23
pytest==7.4.3
python-jose==3.3.0
requests==2.31.0
charset-normalizer>=3.2.0
